mall/requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==4.2.11
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.2
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
oauthlib==3.2.2
packaging==24.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
